Decorate login form
