+++ src/components/Button.jsx
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+const Button = styled.button({
+  backgroundColor: '#4F46E5',
+  borderColor: 'transparent',
+  borderRadius: '.375rem',
+  padding: '.5rem 1rem',
+  opacity: 0.9,
+  fontWeight: 500,
+  color: '#FFFFFF',
+  cursor: 'pointer',
+  '&:hover': {
+    opacity: 1.0,
+  },
+});
+
+export default Button;
