+++ src/pages/RestaurantPage/ReviewForm/ReviewForm.jsx
@@ -1,26 +1,13 @@
 import styled from '@emotion/styled';
 
 import TextField from '@/components/TextField';
+import Button from '@/components/Button';
 
 const Form = styled.div({
   display: 'flex',
   flexWrap: 'wrap',
   gap: '1rem',
   padding: '2rem .5rem',
-});
-
-const Button = styled.button({
-  backgroundColor: '#4F46E5',
-  borderColor: 'transparent',
-  borderRadius: '.375rem',
-  padding: '.5rem 1rem',
-  opacity: 0.9,
-  fontWeight: 500,
-  color: '#FFFFFF',
-  cursor: 'pointer',
-  '&:hover': {
-    opacity: 1.0,
-  },
 });
 
 export default function ReviewForm({ fields, onChange, onSubmit }) {
